fix: ⚡️ move getbatchfromhistory to graphutils
Fixes #59
